+++ xtask/src/lib.rs
@@ -12,6 +12,7 @@ pub(crate) mod formatting;
 pub(crate) mod metadata;
 pub(crate) mod model;
 pub(crate) mod product_package;
+pub(crate) mod release;
 pub(crate) mod reproducibility;
 pub(crate) mod source;
 pub(crate) mod toolchain;